+++ api/src/integrations/slack.rs
@@ -600,8 +600,8 @@ impl NotificationSourceService for SlackService {
         };
 
         // ⚠️ For some reason, the star must be added before being removed
-        // Maybe because it does not exists as a `star` but as `save for later` in the Slack API
-        // Nonetheless, the `stars.remove` method actually remove the `save for later` from the message
+        // Maybe because it does not exists as a `star` but as `saved for later` in the Slack API
+        // Nonetheless, the `stars.remove` method actually remove the `saved for later` from the message
         self.stars_add(&slack_api_token, channel, message, file, file_comment)
             .await?;
         self.stars_remove(&slack_api_token, channel, message, file, file_comment)