+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Display Linear issue new comments
 - Display Linear project on notification item
 - Display Linear Project and Team icons
-- Connect to Slack and receive "save for later" (aka. "stars") events
+- Connect to Slack and receive "saved for later" (aka. "stars") events
 - Add Slack "saved for later" as notifications
 - 2 way sync Slack "saved for later" and Todoist tasks